aider: Add --encoding with default of utf8, and use it in the `open`.
# Aider chat conversation:

USER: add --encoding with default of utf8, and use it in the `open`

ASSISTANT: To add the `--encoding` option with a default of `utf8`, we need to modify the `main` function in the `grep_ast.py` file. We will add a new argument to the `argparse.ArgumentParser()` instance and use this argument when opening the file.

Here is the change:

```python
grep_ast.py
<<<<<<< ORIGINAL
    parser = argparse.ArgumentParser()
    parser.add_argument("-i", "--ignore-case", action="store_true", help="ignore case distinctions")
    parser.add_argument("--no-pretty", action="store_true", help="disable pretty printing")
    parser.add_argument("pat", help="the pattern to search for")
    parser.add_argument("filenames", nargs='+', help="the files to display")
    args = parser.parse_args()

    # Build the AST
    for filename in args.filenames:
        with open(filename, "r") as file:
            code = file.read()
=======
    parser = argparse.ArgumentParser()
    parser.add_argument("-i", "--ignore-case", action="store_true", help="ignore case distinctions")
    parser.add_argument("--no-pretty", action="store_true", help="disable pretty printing")
    parser.add_argument("--
^C KeyboardInterrupt

USER: add --encoding with default of utf8, and use it in the `open`

ASSISTANT: To add the `--encoding` option with a default of `utf8`, we need to modify the `main` function in the `grep_ast.py` file. We will add a new argument to the `argparse.ArgumentParser()` instance and use this argument when opening the file.

Here is the change:

```python
grep_ast.py
<<<<<<< ORIGINAL
    parser = argparse.ArgumentParser()
    parser.add_argument("-i", "--ignore-case", action="store_true", help="ignore case distinctions")
    parser.add_argument("--no-pretty", action="store_true", help="disable pretty printing")
    parser.add_argument("pat", help="the pattern to search for")
    parser.add_argument("filenames", nargs='+', help="the files to display")
    args = parser.parse_args()

    # Build the AST
    for filename in args.filenames:
        with open(filename, "r") as file:
            code = file.read()
=======
    parser = argparse.ArgumentParser()
    parser.add_argument("-i", "--ignore-case", action="store_true", help="ignore case distinctions")
    parser.add_argument("--no-pretty", action="store_true", help="disable pretty printing")
    parser.add_argument("--encoding", default="utf8", help="file encoding")
    parser.add_argument("pat", help="the pattern to search for")
    parser.add_argument("filenames", nargs='+', help="the files to display")
    args = parser.parse_args()

    # Build the AST
    for filename in args.filenames:
        with open(filename, "r", encoding=args.encoding) as file:
            code = file.read()
>>>>>>> UPDATED
```
This change adds a new command line argument `--encoding` with a default value of `utf8`. This encoding is then used when opening the file for reading.